add e.target.value to JSON.stringify
