+++ src/pages/Ratings/Ratings.jsx
@@ -1,39 +1,10 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState, useEffect, Component } from 'react';
 import { FaStar } from 'react-icons/fa'
-
-
-class Review extends Component {
-    state = {
-        invalidForm: true,
-        formData: {
-            content: ''
-        }
-    };
-    formRef = React.createRef();
-
-    handleSubmit = (e) => {
-        e.preventDefault();
-        this.props.handleAddReview(this.state.formData)
-    };
-
-    handleChange = (e) => {
-        const formData = { ... this.state.formData, [e.target.name]: e.target.value };
-        this.setState({
-            formData,
-            invalidForm: !this.formRef.current.checkValidity
-        })
-    }
-
-}
-
-
-
 
 
 const StarRating = () => {
     const [rating, setRating] = useState(null);
     const [hover, setHover] = useState(null);
-
 
     return <div>
         {[...Array(5)].map((star, idx) => {
@@ -47,7 +18,7 @@ const StarRating = () => {
                     onClick={(e) => saveRating(e, setRating)}
                 />
                 <FaStar className="star"
-                    color={ratingValue <= (hover || rating) ? '#ffc107' : '#e4e5e9'} size={50}
+                    color={ratingValue <= (hover || rating) ? '#ffc107' : '#e4e5e9'} size={40}
                     onMouseEnter={() => setHover(ratingValue)}
                     onMouseLeave={() => setHover(null)}
                 />
@@ -65,7 +36,7 @@ const saveRating = (e, setRating) => {
             'Content-Type': 'application/json',
             'Authorization': localStorage.getItem('token')
         },
-        body: JSON.stringify({ rating: 5 })
+        body: JSON.stringify({ rating: e.target.value })
 
     })
         .then(res => res.json())
@@ -73,4 +44,4 @@ const saveRating = (e, setRating) => {
         .catch(err => console.error(err))
 }
 
-export default StarRating
+export default StarRating;
